88a37067ad24 replay gib_eks

+++ saml-core/src/main/java/org/keycloak/saml/processing/core/saml/v2/util/XMLTimeUtil.java
@@ -130,7 +130,7 @@ public class XMLTimeUtil {
      * @return
      */
     public static long inMilis(int valueInMins) {
-        return (long) valueInMins * 60 * 1000;
+        return valueInMins * 60 * 1000;
     }
 
     /**
